+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.29
+* Fixed routePageBuilder usage from BetterPlayerConfiguration
+* Added overflowMenuIcon, playbackSpeedIcon, qualitiesIcon, subtitlesIcon, overflowMenuIconsColor to BetterPlayerControlsConfiguration
+
 ## 0.0.28
 * Fixed subtitles overflow issue when transitioning between fullscreen and normal state
 * Added alignment and backgroundColor in BetterPlayerSubtitlesConfiguration